split AST traversal
