+++ teal/traversal.lua
@@ -1,0 +1,592 @@
+local _tl_compat; if (tonumber((_VERSION or ''):match('[%d.]*$')) or 0) < 5.3 then local p, m = pcall(require, 'compat53.module'); if p then _tl_compat = m end end; local assert = _tl_compat and _tl_compat.assert or assert; local io = _tl_compat and _tl_compat.io or io; local ipairs = _tl_compat and _tl_compat.ipairs or ipairs; local string = _tl_compat and _tl_compat.string or string; local table = _tl_compat and _tl_compat.table or table; local type = type
+
+
+local tldebug = require("teal.debug")
+local TL_DEBUG = tldebug.TL_DEBUG
+local TL_DEBUG_MAXLINE = tldebug.TL_DEBUG_MAXLINE
+
+local types = require("teal.types")
+
+
+
+
+
+
+
+
+
+
+
+
+
+local show_type = types.show_type
+
+local parser = require("teal.parser")
+
+
+
+local traversal = { VisitorCallbacks = {}, Visitor = {} }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+function traversal.fields_of(t, meta)
+   local i = 1
+   local field_order, fields
+   if meta then
+      field_order, fields = t.meta_field_order, t.meta_fields
+   else
+      field_order, fields = t.field_order, t.fields
+   end
+   if not fields then
+      return function()
+      end
+   end
+   return function()
+      local name = field_order[i]
+      if not name then
+         return nil
+      end
+      i = i + 1
+      return name, fields[name]
+   end
+end
+
+local tl_debug_indent = 0
+
+
+
+
+
+
+local tl_debug_entry = nil
+local tl_debug_y = 1
+
+local function tl_debug_loc(y, x)
+   return (tostring(y) or "?") .. ":" .. (tostring(x) or "?")
+end
+
+local function tl_debug_indent_push(mark, y, x, fmt, ...)
+   if tl_debug_entry then
+      if tl_debug_entry.y and (tl_debug_entry.y > tl_debug_y) then
+         io.stderr:write("\n")
+         tl_debug_y = tl_debug_entry.y
+      end
+      io.stderr:write(("   "):rep(tl_debug_indent) .. tl_debug_entry.mark .. " " ..
+      tl_debug_loc(tl_debug_entry.y, tl_debug_entry.x) .. " " ..
+      tl_debug_entry.msg .. "\n")
+      io.stderr:flush()
+      tl_debug_entry = nil
+      tl_debug_indent = tl_debug_indent + 1
+   end
+   tl_debug_entry = {
+      mark = mark,
+      y = y,
+      x = x,
+      msg = fmt:format(...),
+   }
+end
+
+local function tl_debug_indent_pop(mark, single, y, x, fmt, ...)
+   if tl_debug_entry then
+      local msg = tl_debug_entry.msg
+      if fmt then
+         msg = fmt:format(...)
+      end
+      if y and (y > tl_debug_y) then
+         io.stderr:write("\n")
+         tl_debug_y = y
+      end
+      io.stderr:write(("   "):rep(tl_debug_indent) .. single .. " " .. tl_debug_loc(y, x) .. " " .. msg .. "\n")
+      io.stderr:flush()
+      tl_debug_entry = nil
+   else
+      tl_debug_indent = tl_debug_indent - 1
+      if fmt then
+         io.stderr:write(("   "):rep(tl_debug_indent) .. mark .. " " .. fmt:format(...) .. "\n")
+         io.stderr:flush()
+      end
+   end
+end
+
+local recurse_type
+
+local function aggregate_type_walker(s, ast, visit)
+   local xs = {}
+   for i, child in ipairs(ast.types) do
+      xs[i] = recurse_type(s, child, visit)
+   end
+   return xs
+end
+
+local function record_like_type_walker(s, ast, visit)
+   local xs = {}
+   if ast.interface_list then
+      for _, child in ipairs(ast.interface_list) do
+         table.insert(xs, recurse_type(s, child, visit))
+      end
+   end
+   if ast.elements then
+      table.insert(xs, recurse_type(s, ast.elements, visit))
+   end
+   if ast.fields then
+      for _, child in traversal.fields_of(ast) do
+         table.insert(xs, recurse_type(s, child, visit))
+      end
+   end
+   if ast.meta_fields then
+      for _, child in traversal.fields_of(ast, "meta") do
+         table.insert(xs, recurse_type(s, child, visit))
+      end
+   end
+   return xs
+end
+
+local type_walkers = {
+   ["typevar"] = false,
+   ["unresolved_typearg"] = false,
+   ["unresolvable_typearg"] = false,
+   ["self"] = false,
+   ["enum"] = false,
+   ["boolean"] = false,
+   ["string"] = false,
+   ["nil"] = false,
+   ["thread"] = false,
+   ["userdata"] = false,
+   ["number"] = false,
+   ["integer"] = false,
+   ["circular_require"] = false,
+   ["boolean_context"] = false,
+   ["emptytable"] = false,
+   ["unresolved_emptytable_value"] = false,
+   ["any"] = false,
+   ["unknown"] = false,
+   ["invalid"] = false,
+   ["none"] = false,
+   ["*"] = false,
+
+   ["generic"] = function(s, ast, visit)
+      local xs = {}
+      for _, child in ipairs(ast.typeargs) do
+         table.insert(xs, recurse_type(s, child, visit))
+      end
+      table.insert(xs, recurse_type(s, ast.t, visit))
+      return xs
+   end,
+   ["tuple"] = function(s, ast, visit)
+      local xs = {}
+      for i, child in ipairs(ast.tuple) do
+         xs[i] = recurse_type(s, child, visit)
+      end
+      return xs
+   end,
+   ["union"] = aggregate_type_walker,
+   ["tupletable"] = aggregate_type_walker,
+   ["poly"] = aggregate_type_walker,
+   ["map"] = function(s, ast, visit)
+      return {
+         recurse_type(s, ast.keys, visit),
+         recurse_type(s, ast.values, visit),
+      }
+   end,
+   ["record"] = record_like_type_walker,
+   ["interface"] = record_like_type_walker,
+   ["function"] = function(s, ast, visit)
+      local xs = {}
+      if ast.args then
+         for _, child in ipairs(ast.args.tuple) do
+            table.insert(xs, recurse_type(s, child, visit))
+         end
+      end
+      if ast.rets then
+         for _, child in ipairs(ast.rets.tuple) do
+            table.insert(xs, recurse_type(s, child, visit))
+         end
+      end
+      return xs
+   end,
+   ["nominal"] = function(s, ast, visit)
+      local xs = {}
+      if ast.typevals then
+         for _, child in ipairs(ast.typevals) do
+            table.insert(xs, recurse_type(s, child, visit))
+         end
+      end
+      return xs
+   end,
+   ["typearg"] = function(s, ast, visit)
+      return {
+         ast.constraint and recurse_type(s, ast.constraint, visit),
+      }
+   end,
+   ["array"] = function(s, ast, visit)
+      return {
+         recurse_type(s, ast.elements, visit),
+      }
+   end,
+   ["literal_table_item"] = function(s, ast, visit)
+      return {
+         recurse_type(s, ast.ktype, visit),
+         recurse_type(s, ast.vtype, visit),
+      }
+   end,
+   ["typedecl"] = function(s, ast, visit)
+      return {
+         recurse_type(s, ast.def, visit),
+      }
+   end,
+}
+
+recurse_type = function(s, ast, visit)
+   local kind = ast.typename
+
+   if TL_DEBUG then
+      tl_debug_indent_push("---", ast.y, ast.x, "[%s] = %s", kind, show_type(ast))
+   end
+
+   local cbs = visit.cbs
+   local cbkind = cbs and cbs[kind]
+   if cbkind then
+      local cbkind_before = cbkind.before
+      if cbkind_before then
+         cbkind_before(s, ast)
+      end
+   end
+
+   local xs
+   local walker = type_walkers[ast.typename]
+   if not (type(walker) == "boolean") then
+      xs = walker(s, ast, visit)
+   end
+
+   local ret
+   local cbkind_after = cbkind and cbkind.after
+   if cbkind_after then
+      ret = cbkind_after(s, ast, xs)
+   end
+   local visit_after = visit.after
+   if visit_after then
+      ret = visit_after(s, ast, xs, ret)
+   end
+
+   if TL_DEBUG then
+      tl_debug_indent_pop("---", "---", ast.y, ast.x)
+   end
+
+   return ret
+end
+
+local function recurse_typeargs(s, ast, visit_type)
+   if ast.typeargs then
+      for _, typearg in ipairs(ast.typeargs) do
+         recurse_type(s, typearg, visit_type)
+      end
+   end
+end
+
+local function extra_callback(name,
+   s,
+   ast,
+   xs,
+   visit_node)
+   local cbs = visit_node.cbs
+   if not cbs then return end
+   local nbs = cbs[ast.kind]
+   if not nbs then return end
+   local bs = nbs[name]
+   if not bs then return end
+   bs(s, ast, xs)
+end
+
+local no_traverse_nodes = {
+   ["..."] = true,
+   ["nil"] = true,
+   ["cast"] = true,
+   ["goto"] = true,
+   ["break"] = true,
+   ["label"] = true,
+   ["number"] = true,
+   ["pragma"] = true,
+   ["string"] = true,
+   ["boolean"] = true,
+   ["integer"] = true,
+   ["variable"] = true,
+   ["error_node"] = true,
+   ["identifier"] = true,
+   ["type_identifier"] = true,
+}
+
+function traversal.traverse_nodes(s, root,
+   visit_node,
+   visit_type)
+   if not root then
+
+      return
+   end
+
+   local recurse
+
+   local function walk_children(ast, xs)
+      for i, child in ipairs(ast) do
+         xs[i] = recurse(child)
+      end
+   end
+
+   local function walk_vars_exps(ast, xs)
+      xs[1] = recurse(ast.vars)
+      if ast.decltuple then
+         xs[2] = recurse_type(s, ast.decltuple, visit_type)
+      end
+      extra_callback("before_exp", s, ast, xs, visit_node)
+      if ast.exps then
+         xs[3] = recurse(ast.exps)
+      end
+   end
+
+   local function walk_named_function(ast, xs)
+      recurse_typeargs(s, ast, visit_type)
+      xs[1] = recurse(ast.name)
+      xs[2] = recurse(ast.args)
+      xs[3] = recurse_type(s, ast.rets, visit_type)
+      extra_callback("before_statements", s, ast, xs, visit_node)
+      xs[4] = recurse(ast.body)
+   end
+
+   local walkers = {
+      ["op"] = function(ast, xs)
+         xs[1] = recurse(ast.e1)
+         local p1 = ast.e1.op and ast.e1.op.prec or nil
+         if ast.op.op == ":" and ast.e1.kind == "string" then
+            p1 = -999
+         end
+         xs[2] = p1
+         if ast.op.arity == 2 then
+            extra_callback("before_e2", s, ast, xs, visit_node)
+            if ast.op.op == "is" or ast.op.op == "as" then
+               xs[3] = recurse_type(s, ast.e2.casttype, visit_type)
+            else
+               xs[3] = recurse(ast.e2)
+            end
+            xs[4] = (ast.e2.op and ast.e2.op.prec)
+         end
+      end,
+
+      ["statements"] = walk_children,
+      ["argument_list"] = walk_children,
+      ["literal_table"] = walk_children,
+      ["variable_list"] = walk_children,
+      ["expression_list"] = walk_children,
+
+      ["literal_table_item"] = function(ast, xs)
+         xs[1] = recurse(ast.key)
+         xs[2] = recurse(ast.value)
+         if ast.itemtype then
+            xs[3] = recurse_type(s, ast.itemtype, visit_type)
+         end
+      end,
+
+      ["assignment"] = walk_vars_exps,
+      ["local_declaration"] = walk_vars_exps,
+      ["global_declaration"] = walk_vars_exps,
+
+      ["local_type"] = function(ast, xs)
+
+
+         xs[1] = recurse(ast.var)
+         xs[2] = recurse(ast.value)
+      end,
+
+      ["global_type"] = function(ast, xs)
+
+
+         xs[1] = recurse(ast.var)
+         if ast.value then
+            xs[2] = recurse(ast.value)
+         end
+      end,
+
+      ["if"] = function(ast, xs)
+         for _, e in ipairs(ast.if_blocks) do
+            table.insert(xs, recurse(e))
+         end
+      end,
+
+      ["if_block"] = function(ast, xs)
+         if ast.exp then
+            xs[1] = recurse(ast.exp)
+         end
+         extra_callback("before_statements", s, ast, xs, visit_node)
+         xs[2] = recurse(ast.body)
+      end,
+
+      ["while"] = function(ast, xs)
+         xs[1] = recurse(ast.exp)
+         extra_callback("before_statements", s, ast, xs, visit_node)
+         xs[2] = recurse(ast.body)
+      end,
+
+      ["repeat"] = function(ast, xs)
+         xs[1] = recurse(ast.body)
+         xs[2] = recurse(ast.exp)
+      end,
+
+      ["macroexp"] = function(ast, xs)
+         recurse_typeargs(s, ast, visit_type)
+         xs[1] = recurse(ast.args)
+         xs[2] = recurse_type(s, ast.rets, visit_type)
+         extra_callback("before_exp", s, ast, xs, visit_node)
+         xs[3] = recurse(ast.exp)
+      end,
+
+      ["function"] = function(ast, xs)
+         recurse_typeargs(s, ast, visit_type)
+         xs[1] = recurse(ast.args)
+         xs[2] = recurse_type(s, ast.rets, visit_type)
+         extra_callback("before_statements", s, ast, xs, visit_node)
+         xs[3] = recurse(ast.body)
+      end,
+      ["local_function"] = walk_named_function,
+      ["global_function"] = walk_named_function,
+      ["record_function"] = function(ast, xs)
+         recurse_typeargs(s, ast, visit_type)
+         xs[1] = recurse(ast.fn_owner)
+         xs[2] = recurse(ast.name)
+         extra_callback("before_arguments", s, ast, xs, visit_node)
+         xs[3] = recurse(ast.args)
+         xs[4] = recurse_type(s, ast.rets, visit_type)
+         extra_callback("before_statements", s, ast, xs, visit_node)
+         xs[5] = recurse(ast.body)
+      end,
+      ["local_macroexp"] = function(ast, xs)
+
+         xs[1] = recurse(ast.name)
+         xs[2] = recurse(ast.macrodef.args)
+         xs[3] = recurse_type(s, ast.macrodef.rets, visit_type)
+         extra_callback("before_exp", s, ast, xs, visit_node)
+         xs[4] = recurse(ast.macrodef.exp)
+      end,
+
+      ["forin"] = function(ast, xs)
+         xs[1] = recurse(ast.vars)
+         xs[2] = recurse(ast.exps)
+         extra_callback("before_statements", s, ast, xs, visit_node)
+         xs[3] = recurse(ast.body)
+      end,
+
+      ["fornum"] = function(ast, xs)
+         xs[1] = recurse(ast.var)
+         xs[2] = recurse(ast.from)
+         xs[3] = recurse(ast.to)
+         xs[4] = ast.step and recurse(ast.step)
+         extra_callback("before_statements", s, ast, xs, visit_node)
+         xs[5] = recurse(ast.body)
+      end,
+
+      ["return"] = function(ast, xs)
+         xs[1] = recurse(ast.exps)
+      end,
+
+      ["do"] = function(ast, xs)
+         xs[1] = recurse(ast.body)
+      end,
+
+      ["paren"] = function(ast, xs)
+         xs[1] = recurse(ast.e1)
+      end,
+
+      ["newtype"] = function(ast, xs)
+         xs[1] = recurse_type(s, ast.newtype, visit_type)
+      end,
+
+      ["argument"] = function(ast, xs)
+         if ast.argtype then
+            xs[1] = recurse_type(s, ast.argtype, visit_type)
+         end
+      end,
+   }
+
+   if not visit_node.allow_missing_cbs and not visit_node.cbs then
+      error("missing cbs in visit_node")
+   end
+   local visit_after = visit_node.after
+
+   recurse = function(ast)
+      local xs = {}
+      local kind = assert(ast.kind)
+      local kprint
+
+      local cbs = visit_node.cbs
+      local cbkind = cbs and cbs[kind]
+      if cbkind then
+         if cbkind.before then
+            cbkind.before(s, ast)
+         end
+      end
+
+      if TL_DEBUG then
+         if ast.y > TL_DEBUG_MAXLINE then
+            error("Halting execution at input line " .. ast.y)
+         end
+         kprint = kind == "op" and "op " .. ast.op.op or
+         kind == "identifier" and "identifier " .. ast.tk or
+         kind
+         tl_debug_indent_push("{{{", ast.y, ast.x, "[%s]", kprint)
+      end
+
+      local fn = walkers[kind]
+      if fn then
+         fn(ast, xs)
+      else
+         assert(no_traverse_nodes[kind])
+      end
+
+      local ret
+      local cbkind_after = cbkind and cbkind.after
+      if cbkind_after then
+         ret = cbkind_after(s, ast, xs)
+      end
+      if visit_after then
+         ret = visit_after(s, ast, xs, ret)
+      end
+
+      if TL_DEBUG then
+         local typ = ast.debug_type and " = " .. show_type(ast.debug_type) or ""
+         tl_debug_indent_pop("}}}", "***", ast.y, ast.x, "[%s]%s", kprint, typ)
+      end
+
+      return ret
+   end
+
+   return recurse(root)
+end
+
+return traversal
